Улучшил код еще лучше
